scopesnap-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-multipart==0.0.12
pillow==10.4.0
weasyprint==62.3
httpx==0.27.2
pydantic-settings==2.5.2
google-generativeai==0.8.3
stripe==10.12.0
resend==2.4.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
svix==1.90.0

# AI Cascade dependencies
# ultralytics/PyTorch (~2GB) excluded — ONNX Runtime used for inference instead
onnxruntime==1.18.1       # YOLO inference via ONNX — 50MB, no PyTorch needed
xgboost>=2.0.0            # Sensor fault detection model (90.09% accuracy)
joblib>=1.3.0             # Loading .pkl model files
numpy>=1.24.0             # Feature engineering for sensor model + ONNX preprocessing
sentry-sdk[fastapi]==2.19.2
slowapi==0.1.9            # Rate limiting for Gemini/AI endpoints
openpyxl>=3.1.0              # WS-A: read SnapAI_HVAC_Master_Price_List_2026.xlsx in load_repo.py
